In English, due to the probable future audience. This Makefile revisions makes it possible to create two different targets, make TARGET=sloppy make TARGET=strict The sloppy one accepts Per-ip and Perip, the strict one only Perip. Thus, the speller project should compile with the strict target.
In order to achieve this we define the source files in the beginning, in SRCS.
New source files should be added here.

Then we have two conditional loops, one for strict and one for sloppy.
Both have ikal.fst as their final goal.

Then it is pretty much business as usual, until a new conditional loop:
ifeq ($(TARGET)=strict)
... we remove all lines containing SUB in the source files
and if target equals sloppy, we do not.

Thus, non-standard forms may be included in the lexicon with !SUB marks.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/st/kal@8183 c7155fb1-f0a7-4240-a2fc-2600b6f42f90