Copied most KAL source files from old infra to new. Many of the files had been copied outside svn earlier, and had lost their history. Now the history is restored. Other files have been modified, they now contain the latest changes in the old infra.
NB! One file remains: st/kal/src/kal-lex.txt. This file (or the splitted part of it) has been modified both in the new and old infra, and merging the changes requires more work. To be done soon.

The work is done in preparation for moving all KAL work to the new infra.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/kal@72663 c7155fb1-f0a7-4240-a2fc-2600b6f42f90