Finally managed to get rid of the following warning:
/usr/local/bin/hfst-compose-intersect: warning: 
Found output multi-char symbols ("AA") in 
transducer in file morphology/kal.lexc.hfst which are not found on the
output tape of transducers in file phonology/kal-phon.hfst.

(and a similar message for TA after AA was cleaned).

The problem was that both AA and TA are declared as multichar symbols related to prefixes, and that the propernouns.lexc file contained acronyms with these exact letter sequences in them. Those letter sequences were interpreted as the multichar symbols, which of course showed up on the surface side of the fst.

The solution was to replace the regular lexc entries with regex entries in which each letter was entered as a separate symbol. This forces lexc to read each letter as a separate symbol, and we thus get rid of the multichar interpretation.

The correct solution is to change the multichar symbols into something that can not be interpreted as part of regular words and acronyms, but since the KAL lexicon is way beyond my knowledge, I didn't dare going that route. The chosen solution is working very well, and does not change the semantics of the morphology in any way.